minor restrucutring
