+++ apps/client/src/views/Authentication/ResetPassword.tsx
@@ -1,5 +1,5 @@
 import { Label } from '@radix-ui/react-label';
-import React, { useState, FC, FormEvent, useEffect } from 'react';
+import React, { useState, FC, FormEvent } from 'react';
 import { useMutation, useQuery } from '@tanstack/react-query';
 import { Navigate, useNavigate, useParams } from 'react-router-dom';
 import {
